move all the duckdb stuff out of `+layout.js`
