+++ packages/evidence/package.json
@@ -38,6 +38,7 @@
 	},
 	"dependencies": {
 		"@evidence-dev/preprocess": "workspace:^",
+        "@evidence-dev/universal-sql": "workspace:^",
 		"@evidence-dev/plugin-connector": "workspace:^",
 		"@evidence-dev/telemetry": "workspace:^",
 		"@sveltejs/adapter-static": "1.0.0",
