+++ packages/universal-sql/package.json
@@ -16,5 +16,16 @@
 		"chalk": "^5.2.0",
 		"duckdb": "^0.8.1",
 		"parquet-wasm": "0.3.1"
-	}
+	},
+    "exports": {
+        ".": {
+            "default": "./src/index.js",
+            "types": "./src/index.d.ts"
+        },
+        "./client-duckdb": {
+            "browser": "./src/client-duckdb/browser.js",
+            "node": "./src/client-duckdb/node.js",
+            "types": "./src/client-duckdb/browser.d.ts"
+        }
+    }
 }
