+++ sites/example-project/package.json
@@ -10,8 +10,9 @@
 	"dependencies": {
 		"@duckdb/duckdb-wasm": "^1.27.0",
 		"@evidence-dev/component-utilities": "workspace:^",
+        "@evidence-dev/universal-sql": "workspace:^",
 		"@evidence-dev/core-components": "workspace:^",
-		"@evidence-dev/duckdb": "workspace:^0.1.0",
+		"@evidence-dev/duckdb": "workspace:^",
 		"@evidence-dev/plugin-connector": "workspace:^",
 		"@evidence-dev/telemetry": "1.0.5",
 		"@sveltejs/kit": "1.21.0",
